Split out default shortcuts from merging custom shortcuts function.
